requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
captum==0.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.6.0
configparser==5.2.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
deepdish==0.3.7
defusedxml==0.7.1
docker-pycreds==0.4.0
ecco==0.1.2
entrypoints==0.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.3
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.17.3
huggingface-hub==0.4.0
hypertools==0.7.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.1
llvmlite==0.36.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numba==0.53.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.2.52
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
ppca==0.0.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pyglet==1.5.0
Pygments==2.11.2
pynndescent==0.5.6
pyparsing==3.0.8
pyrsistent==0.18.0
pytest==6.2.2
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==0.24.2
scipy==1.5.2
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.9
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tables==3.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.10.1
torchmetrics==0.7.3
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
transformers==4.11.3
typing_extensions==4.1.1
umap-learn==0.5.2
urllib3==1.26.9
wandb==0.12.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
yarl==1.7.2
zipp==3.6.0
